Refactor, and add NCML to JSON
